rgw: add usage logging support in Requester Pays bucket requests
If the request in one Requester Pays bucket succeeds the usage logging is
charged to the requester instead of the bucket owner. If the request was not
successful the bucket owner usage logging is charged.

The new usage logging indexes by 'user, owner and bucket' instead of 'user and
owner' only.

Fixes: #

Signed-off-by: Javier M. Mellid <jmunhoz@igalia.com>